refactoring DivisionsList y DivisionsListCredit
