+++ src/components/group/divisions/DivisionListCredit.jsx
@@ -1,48 +1,12 @@
-import { useEffect, useMemo, useState } from 'react'
+import { useMemo, useState } from 'react'
 import { calculateFinalResultCredit } from '../../../logic/logic'
 import { useGroupStore } from '../../../store/store'
 import Card from '../../common/Card'
-import MoneyAmount from '../../common/MoneyAmount'
 
 import { translatePaymentKey } from '../../../utils/utils'
 import { ChevronLeftBtn, ChevronRightBtn } from '../../common/ChevronBtn/chevronBtn'
-import { ArrowIcon } from '../../icons/icons'
-import { PeopleDivisionIllustration } from '../../illustrations/Illustrations'
-
-const EmptyDivisionsList = () => {
-    return (
-        <div className='flex flex-col justify-center items-center p-10'>
-            <PeopleDivisionIllustration width={160} height={180} />
-            <h4 className='text-center'>Aún no hay divisiones con crédito...</h4>
-            <p className='text-center'>
-                {' '}
-                Aquí verás cuánto le corresponde pagar a cada integrante por cada mes
-            </p>
-        </div>
-    )
-}
-
-const DivisionListItem = ({ division }) => {
-    const { personaFrom, personaTo, cantidad } = division
-    return (
-        <li className='rounded mb-3 hover:bg-gray-50 p-4 dark:hover:bg-slate-600'>
-            <div className='flex gap-6 items-center '>
-                <div className='flex gap-5 items-center'>
-                    <div className='flex flex-col gap-1'>
-                        <h5 className='m-0 capitalize w-20 '>{personaFrom.name}</h5>
-                        <div className='flex items-center gap-1'>
-                            <ArrowIcon />
-                            <h5 className='m-0 capitalize ml-auto text-gray-400'>{personaTo.name}</h5>
-                        </div>
-                    </div>
-                </div>
-                <div className='ml-auto'>
-                    <MoneyAmount amount={cantidad} className='bg-primary  text-white font-bold' />
-                </div>
-            </div>
-        </li>
-    )
-}
+import DivisionListItem from './DivisionListItem'
+import EmptyDivisionsList from './EmptyDivisionsList'
 
 const PaymentNavigation = ({ payment, changePayment, creditPayments }) => {
     return (
@@ -82,13 +46,13 @@ const DivisionListCredit = () => {
 
     console.log(memoizedResultsCredit)
 
-    useEffect(() => {
-        console.log('Render DivisionListCredit...')
-    })
     return (
         <Card className='animate-fade'>
             {creditPayments.length === 0 ? (
-                <EmptyDivisionsList />
+                <EmptyDivisionsList
+                    title='Aún no hay divisiones con crédito...'
+                    subtitle=' Aquí verás cuánto le corresponde pagar a cada integrante por cada mes'
+                />
             ) : (
                 <>
                     <PaymentNavigation
